+++ ext/ManoptJuMPExt.jl
@@ -7,41 +7,132 @@ using ManifoldsBase
 using ManifoldDiff
 const MOI = JuMP.MOI
 
-function __init__()
-    setglobal!(Manopt, :JuMP_Optimizer, Optimizer)
-    setglobal!(Manopt, :JuMP_VectorizedManifold, VectorizedManifold)
-    setglobal!(Manopt, :JuMP_ArrayShape, ArrayShape)
-    return nothing
+"""
+    ManoptOptimizer <: MOI.AbstractOptimizer
+
+Represent a solver from `Manopt.jl` within the [`MathOptInterface` (MOI)](@extref JuMP :std:label:`The-MOI-interface`) framework of [`JuMP.jl`](@extref JuMP :std:doc:`index`)
+
+# Fields
+* `problem::`[`AbstractManoptProblem`](@ref) a problem in manopt, especially
+    containing the manifold and the objective function. It can be constructed as soon as
+    the manifold and the objective are present.
+* `manifold::`[`AbstractManifold`](@extref `ManifoldsBase.AbstractManifold`) the manifold on which the optimization is performed.
+* `objective::`[`AbstractManifoldObjective`](@ref) the objective function to be optimized.
+* `state::`[`AbstractManoptSolverState`](@ref) the state specifying the solver to use.
+* `variable_primal_start::Vector{Union{Nothing,Float64}}` starting value for the solver,
+    in a vectorized form that [`JuMP.jl`](@extref JuMP :std:doc:`index`) requires.
+* `sense::`[`MOI.OptimizationSense`](@extref JuMP :jl:type:`MathOptInterface.OptimizationSense`) the sense of optimization,
+  currently only minimization and maximization are supported.
+* `options::Dict{String,Any}`: parameters specifying a solver before the `state`
+  is initialized, so especially which [`AbstractManoptSolverState`](@ref) to use,
+  when setting up the `state.
+All types in brackets can also be `Nothing`, indicating they were not yet initialized.
+"""
+mutable struct ManoptOptimizer <: MOI.AbstractOptimizer
+    problem::Union{Nothing,Manopt.AbstractManoptProblem}
+    manifold::Union{Nothing,ManifoldsBase.AbstractManifold}
+    objective::Union{Nothing,Manopt.AbstractManifoldObjective}
+    state::Union{Nothing,Manopt.AbstractManoptSolverState}
+    # Does this make sense to be elementwise Nothing? On a manifold a partial init is not possible
+    variable_primal_start::Vector{Union{Nothing,Float64}}
+    sense::MOI.OptimizationSense
+    # Not sure what these are for? All parameters set should be reflected in the `state` parameter.
+    options::Dict{String,Any}
+    function ManoptOptimizer()
+        return new(
+            nothing,
+            nothing,
+            nothing,
+            nothing,
+            Union{Nothing,Float64}[],
+            MOI.FEASIBILITY_SENSE,
+            Dict{String,Any}(DESCENT_STATE_TYPE => Manopt.GradientDescentState),
+        )
+    end
+end
+"""
+    Manopt.JuMP_Optimizer()
+
+Represent a solver from `Manopt.jl` within the [`MathOptInterface` (MOI)](@extref JuMP :std:label:`The-MOI-interface`) framework.
+See [`ManoptOptimizer`](@ref) for the fields and their meaning.
+"""
+function Manopt.JuMP_Optimizer(args...)
+    return ManoptOptimizer(args...)
 end
 
-struct VectorizedManifold{M<:ManifoldsBase.AbstractManifold} <: MOI.AbstractVectorSet
+"""
+    ManifoldSet{M<:ManifoldsBase.AbstractManifold} <: MOI.AbstractVectorSet
+
+Model a manifold from [`ManifoldsBase.jl`](@extref) as a vectorial set in the
+[`MathOptInterface` (MOI)](@extref JuMP :std:label:`The-MOI-interface`).
+This is a slight misuse of notation, since the manifold itself might not be embedded,
+but just be parametrized in a certain way.
+
+# Fields
+
+* `manifold::M`: The manifold in which the variables are constrained to lie.
+  This is a [`ManifoldsBase.AbstractManifold`](@extref) object.
+"""
+struct ManifoldSet{M<:ManifoldsBase.AbstractManifold} <: MOI.AbstractVectorSet
     manifold::M
 end
 
 """
-    MOI.dimension(set::VectorizedManifold)
+    MOI.dimension(set::ManifoldSet)
 
 Return the representation side of points on the (vectorized in representation) manifold.
 As the MOI variables are real, this means if the [`representation_size`](@extref `ManifoldsBase.representation_size-Tuple{AbstractManifold}`)
 yields (in product) `n`, this refers to the vectorized point / tangent vector  from (a subset of ``ℝ^n``).
 """
-function MOI.dimension(set::VectorizedManifold)
+function MOI.dimension(set::ManifoldSet)
     return prod(ManifoldsBase.representation_size(set.manifold))
 end
 
+@doc """
+    RiemannianFunction{MO<:Manopt.AbstractManifoldObjective} <: MOI.AbstractScalarFunction
+A wrapper for a [`AbstractManifoldObjective`](@ref) that can be used
+as a [`MOI.AbstractScalarFunction`](@extref JuMP :jl:type:`MathOptInterface.AbstractScalarFunction`).
+
+
+
+# Fields
+* `func::MO`: The [`AbstractManifoldObjective`](@ref) function to be wrapped.
+"""
 struct RiemannianFunction{MO<:Manopt.AbstractManifoldObjective} <:
        MOI.AbstractScalarFunction
     func::MO
 end
 
+@doc """
+    JuMP.jump_function_type(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})
+
+The [`JuMP.jl`](@extref JuMP :std:doc:`index`) function type of a function of type [`RiemannianFunction`](@ref) for any [`AbstractModel`](@extref JuMP.AbstractModel)
+is that function type itself
+"""
 function JuMP.jump_function_type(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})
     return F
 end
 
+@doc """
+    JuMP.jump_function(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})
+
+The [`JuMP.jl`](@extref JuMP :std:doc:`index`) function of a [`RiemannianFunction`](@ref) for any [`AbstractModel`](@extref JuMP.AbstractModel)
+is that function itself.
+"""
 JuMP.jump_function(::JuMP.AbstractModel, f::RiemannianFunction) = f
 
+#
+# The string representation
+# maybe not document this since it seems to be mainly for display reasons
 JuMP.function_string(mime::MIME, f::RiemannianFunction) = string(f.func)
 
+"""
+    MOI.Utilities.map_indices(index_map::Function, func::RiemannianFunction)
+
+The original docstring states something about substituting some variable indices
+by their index map variants.
+On a [`RiemannianFunction`](@ref) there is nothing to substitute,
+"""
 MOI.Utilities.map_indices(::Function, func::RiemannianFunction) = func
 
 # We we don't support `MOI.modify` and `RiemannianFunction` is not mutable, no need to copy anything
@@ -50,51 +141,29 @@ Base.copy(func::RiemannianFunction) = func
 # This is called for instance when the user does `@objective(model, Min, func)`.
 # JuMP only accepts subtypes of `MOI.AbstractFunction` as objective so we wrap `func`.
 # It will then be allowed to go through all the MOI layers because it is of the right type
-# We will then receive it in `MOI.set(::Optimizer, ::MOI.ObjectiveFunction, RiemannianFunction)`
+# We will then receive it in `MOI.set(::ManoptOptimizer, ::MOI.ObjectiveFunction, RiemannianFunction)`
 # where we will unwrap it and recover `func`.
+@doc """
+    JuMP.set_objective_function(model::JuMP.Model, obj::Manopt.AbstractManifoldObjective)
+
+Set the objective function of a [`JuMP.Model`](@extref) `model` to an [`AbstractManifoldObjective`](@ref) `obj`.
+This allows to use `@objective` with an objective from `Manopt.jl`.
+"""
 function JuMP.set_objective_function(
     model::JuMP.Model, func::Manopt.AbstractManifoldObjective
 )
     return JuMP.set_objective_function(model, RiemannianFunction(func))
 end
 
-mutable struct Optimizer <: MOI.AbstractOptimizer
-    # Manifold in which all the decision variables leave
-    manifold::Union{Nothing,ManifoldsBase.AbstractManifold}
-    # Description of the problem in Manopt
-    problem::Union{Nothing,Manopt.AbstractManoptProblem}
-    # State of the optimizer
-    state::Union{Nothing,Manopt.AbstractManoptSolverState}
-    # Starting value for each variable
-    variable_primal_start::Vector{Union{Nothing,Float64}}
-    # Sense of the optimization, that is whether it is for example min, max or no objective
-    sense::MOI.OptimizationSense
-    # Objective function of the optimization
-    objective::Union{Nothing,Manopt.AbstractManifoldObjective}
-    # Solver parameters set with `MOI.RawOptimizerAttribute`
-    options::Dict{String,Any}
-    function Optimizer()
-        return new(
-            nothing,
-            nothing,
-            nothing,
-            Union{Nothing,Float64}[],
-            MOI.FEASIBILITY_SENSE,
-            nothing,
-            Dict{String,Any}(DESCENT_STATE_TYPE => Manopt.GradientDescentState),
-        )
-    end
-end
-
 """
-    MOI.get(::Optimizer, ::MOI.SolverVersion)
+    MOI.get(::ManoptOptimizer, ::MOI.SolverVersion)
 
 Return the version of the Manopt solver, it corresponds to the version of
 Manopt.jl.
 """
-MOI.get(::Optimizer, ::MOI.SolverVersion) = "0.4.37"
+MOI.get(::ManoptOptimizer, ::MOI.SolverVersion) = "Manopt.jl $(pkgversion(Manopt))"
 
-function MOI.is_empty(model::Optimizer)
+function MOI.is_empty(model::ManoptOptimizer)
     return isnothing(model.manifold) &&
            isempty(model.variable_primal_start) &&
            isnothing(model.objective) &&
@@ -102,11 +171,11 @@ function MOI.is_empty(model::Optimizer)
 end
 
 """
-    MOI.empty!(model::ManoptJuMPExt.Optimizer)
+    MOI.empty!(model::ManoptOptimizer)
 
 Clear all model data from `model` but keep the `options` set.
 """
-function MOI.empty!(model::Optimizer)
+function MOI.empty!(model::ManoptOptimizer)
     model.manifold = nothing
     model.problem = nothing
     model.state = nothing
@@ -117,12 +186,12 @@ function MOI.empty!(model::Optimizer)
 end
 
 """
-    MOI.supports(::Optimizer, attr::MOI.RawOptimizerAttribute)
+    MOI.supports(::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)
 
 Return a `Bool` indicating whether `attr.name` is a valid option name
 for `Manopt`.
 """
-function MOI.supports(::Optimizer, ::MOI.RawOptimizerAttribute)
+function MOI.supports(::ManoptOptimizer, ::MOI.RawOptimizerAttribute)
     # FIXME Ideally, this should only return `true` if it is a valid keyword argument for
     #       one of the `...DescentState()` constructors. Is there an easy way to check this ?
     #       Does it depend on the different solvers ?
@@ -130,76 +199,75 @@ function MOI.supports(::Optimizer, ::MOI.RawOptimizerAttribute)
 end
 
 """
-    MOI.get(model::Optimizer, attr::MOI.RawOptimizerAttribute)
+    MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)
 
-Return last `value` set by `MOI.set(model, attr, value)`.
+Return last `value` set by [`set`](@extref `MathOptInterface.set`)`(model, attr, value)`.
 """
-function MOI.get(model::Optimizer, attr::MOI.RawOptimizerAttribute)
+function MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)
     return model.options[attr.name]
 end
 
 """
-    MOI.get(model::Optimizer, attr::MOI.RawOptimizerAttribute)
+    MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)
 
 Set the value for the keyword argument `attr.name` to give for the constructor
 `model.options[DESCENT_STATE_TYPE]`.
 """
-function MOI.set(model::Optimizer, attr::MOI.RawOptimizerAttribute, value)
+function MOI.set(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute, value)
     model.options[attr.name] = value
     return nothing
 end
 
 """
-    MOI.get(::Optimizer, ::MOI.SolverName)
+    MOI.get(::ManoptOptimizer, ::MOI.SolverName)
 
-Return the name of the `Optimizer` with the value of
+Return the name of the [`ManoptOptimizer`](@ref) with the value of
 the `descent_state_type` option.
 """
-function MOI.get(model::Optimizer, ::MOI.SolverName)
-    return "Manopt with $(model.options[DESCENT_STATE_TYPE])"
+function MOI.get(model::ManoptOptimizer, ::MOI.SolverName)
+    return "A Manopt.jl solver, namely $(model.options[DESCENT_STATE_TYPE])"
 end
 
 """
-    MOI.supports_incremental_interface(::JuMP_Optimizer)
+    MOI.supports_incremental_interface(::ManoptOptimizer)
 
-Return `true` indicating that `Manopt.JuMP_Optimizer` implements
-`MOI.add_constrained_variables` and `MOI.set` for
-`MOI.ObjectiveFunction` so it can be used with [`JuMP.direct_model`](https://jump.dev/JuMP.jl/stable/api/JuMP/#JuMP.direct_model)
-and does not require a `MOI.Utilities.CachingOptimizer`.
-See [`MOI.supports_incremental_interface`](https://jump.dev/JuMP.jl/stable/moi/reference/models/#MathOptInterface.supports_incremental_interface).
+Return `true` indicating that [`ManoptOptimizer`](@ref) implements
+[`add_constrained_variables`](@extref `MathOptInterface.add_constrained_variables`) and [`set`](@extref `MathOptInterface.set`) for
+[`ObjectiveFunction`](@extref `MathOptInterface.ObjectiveFunction`) so it can be used with [`direct_model`](@extref `JuMP.direct_model`)
+and does not require a [`CachingOptimizer`](@extref `MathOptInterface.Utilities.CachingOptimizer`).
+See See [`supports_incremental_interface`](@extref `MathOptInterface.supports_incremental_interface`).
 """
-MOI.supports_incremental_interface(::Optimizer) = true
+MOI.supports_incremental_interface(::ManoptOptimizer) = true
 
 """
-    MOI.copy_to(dest::Optimizer, src::MOI.ModelLike)
+    MOI.copy_to(dest::ManoptOptimizer, src::MOI.ModelLike)
 
-Because `supports_incremental_interface(dest)` is `true`, this simply
-uses `MOI.Utilities.default_copy_to` and copies the variables with
-`MOI.add_constrained_variables` and the objective sense with `MOI.set`.
+Because [`supports_incremental_interface`](@extref `MathOptInterface.supports_incremental_interface`)`(dest)` is `true`, this simply
+uses [`default_copy_to`](@extref `MathOptInterface.Utilities.default_copy_to`) and copies the variables with
+[`add_constrained_variables`](@extref `MathOptInterface.add_constrained_variables`) and the objective sense with [`set`](@extref `MathOptInterface.set`).
 """
-function MOI.copy_to(dest::Optimizer, src::MOI.ModelLike)
+function MOI.copy_to(dest::ManoptOptimizer, src::MOI.ModelLike)
     return MOI.Utilities.default_copy_to(dest, src)
 end
 
 """
-    MOI.supports_add_constrained_variables(::JuMP_Optimizer, ::Type{<:VectorizedManifold})
+    MOI.supports_add_constrained_variables(::ManoptOptimizer, ::Type{<:ManifoldSet})
 
-Return `true` indicating that `Manopt.JuMP_Optimizer` support optimization on
-variables constrained to belong in a vectorized manifold [`Manopt.JuMP_VectorizedManifold`](@ref).
+Return `true` indicating that [`ManoptOptimizer`](@ref) support optimization on
+variables constrained to belong in a vectorized manifold.
 """
-function MOI.supports_add_constrained_variables(::Optimizer, ::Type{<:VectorizedManifold})
+function MOI.supports_add_constrained_variables(::ManoptOptimizer, ::Type{<:ManifoldSet})
     return true
 end
 
 """
-    MOI.add_constrained_variables(model::Optimizer, set::VectorizedManifold)
+    MOI.add_constrained_variables(model::ManoptOptimizer, set::ManifoldSet)
 
-Add `MOI.dimension(set)` variables constrained in `set` and return the list
+Add [`dimension`](@extref `MathOptInterface.dimension`)`(set)` variables constrained in `set` and return the list
 of variable indices that can be used to reference them as well a constraint
-index for the constraint enforcing the membership of the variables in the
-[`Manopt.JuMP_VectorizedManifold`](@ref) `set`.
+index for the constraint enforcing the membership of the variables manifold as a set.
 """
-function MOI.add_constrained_variables(model::Optimizer, set::VectorizedManifold)
+function MOI.add_constrained_variables(model::ManoptOptimizer, set::ManifoldSet)
     F = MOI.VectorOfVariables
     if !isnothing(model.manifold)
         throw(
@@ -220,42 +288,44 @@ function MOI.add_constrained_variables(model::Optimizer, set::VectorizedManifold
 end
 
 """
-    MOI.is_valid(model::Optimizer, vi::MOI.VariableIndex)
+    MOI.is_valid(model::ManoptOptimizer, vi::MOI.VariableIndex)
 
 Return whether `vi` is a valid variable index.
 """
-function MOI.is_valid(model::Optimizer, vi::MOI.VariableIndex)
+function MOI.is_valid(model::ManoptOptimizer, vi::MOI.VariableIndex)
     return !isnothing(model.manifold) &&
-           1 <= vi.value <= MOI.dimension(VectorizedManifold(model.manifold))
+           1 <= vi.value <= MOI.dimension(ManifoldSet(model.manifold))
 end
 
 """
-    MOI.get(model::Optimizer, ::MOI.NumberOfVariables)
+    MOI.get(model::ManoptOptimizer, ::MOI.NumberOfVariables)
 
 Return the number of variables added in the model, this corresponds
-to the [`MOI.dimension`](@ref) of the [`Manopt.JuMP_VectorizedManifold`](@ref).
+to the [`dimension`](@extref JuMP :jl:function:`MathOptInterface.dimension`) of the [`ManifoldSet`](@ref).
 """
-function MOI.get(model::Optimizer, ::MOI.NumberOfVariables)
+function MOI.get(model::ManoptOptimizer, ::MOI.NumberOfVariables)
     if isnothing(model.manifold)
         return 0
     else
-        return MOI.dimension(VectorizedManifold(model.manifold))
+        return MOI.dimension(ManifoldSet(model.manifold))
     end
 end
 
 """
-    MOI.supports(::Manopt.JuMP_Optimizer, attr::MOI.RawOptimizerAttribute)
+    MOI.supports(::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)
 
-Return `true` indicating that `Manopt.JuMP_Optimizer` supports starting values
+Return `true` indicating that [`ManoptOptimizer`](@ref) supports starting values
 for the variables.
 """
-function MOI.supports(::Optimizer, ::MOI.VariablePrimalStart, ::Type{MOI.VariableIndex})
+function MOI.supports(
+    ::ManoptOptimizer, ::MOI.VariablePrimalStart, ::Type{MOI.VariableIndex}
+)
     return true
 end
 
 """
     function MOI.set(
-        model::Optimizer,
+        model::ManoptOptimizer,
         ::MOI.VariablePrimalStart,
         vi::MOI.VariableIndex,
         value::Union{Real,Nothing},
@@ -266,7 +336,7 @@ Set the starting value of the variable of index `vi` to `value`. Note that if
 and hence `MOI.optimize!` unless another starting value is set.
 """
 function MOI.set(
-    model::Optimizer,
+    model::ManoptOptimizer,
     ::MOI.VariablePrimalStart,
     vi::MOI.VariableIndex,
     value::Union{Real,Nothing},
@@ -278,41 +348,41 @@ function MOI.set(
 end
 
 """
-    MOI.supports(::Optimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})
+    MOI.supports(::ManoptOptimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})
 
 Return `true` indicating that `Optimizer` supports being set the objective
 sense (that is, min, max or feasibility) and the objective function.
 """
-function MOI.supports(::Optimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})
+function MOI.supports(::ManoptOptimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})
     return true
 end
 
 """
-    MOI.set(model::Optimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
+    MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
 
-Modify the objective sense to either `MOI.MAX_SENSE`, `MOI.MIN_SENSE` or
-`MOI.FEASIBILITY_SENSE`.
+Modify the objective sense to either [`MAX_SENSE`](@extref), [`MIN_SENSE`](@extref) or
+[`FEASIBILITY_SENSE`](@extref).
 """
-function MOI.set(model::Optimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
+function MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)
     model.sense = sense
     return nothing
 end
 
 """
-    MOI.get(model::Optimizer, ::MOI.ObjectiveSense)
+    MOI.get(model::ManoptOptimizer, ::MOI.ObjectiveSense)
 
-Return the objective sense, defaults to `MOI.FEASIBILITY_SENSE` if no sense has
+Return the objective sense, defaults to [`FEASIBILITY_SENSE`](@extref) if no sense has
 already been set.
 """
-MOI.get(model::Optimizer, ::MOI.ObjectiveSense) = model.sense
+MOI.get(model::ManoptOptimizer, ::MOI.ObjectiveSense) = model.sense
 
 """
-    MOI.set(model::Optimizer, ::MOI.ObjectiveFunction{F}, func::F) where {F}
+    MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveFunction{F}, func::F) where {F}
 
 Set the objective function as `func` for `model`.
 """
 function MOI.set(
-    model::Optimizer, attr::MOI.ObjectiveFunction, func::MOI.AbstractScalarFunction
+    model::ManoptOptimizer, attr::MOI.ObjectiveFunction, func::MOI.AbstractScalarFunction
 )
     backend = MOI.Nonlinear.SparseReverseMode()
     vars = [MOI.VariableIndex(i) for i in eachindex(model.variable_primal_start)]
@@ -338,7 +408,7 @@ function MOI.set(
     return nothing
 end
 
-function MOI.set(model::Optimizer, ::MOI.ObjectiveFunction, func::RiemannianFunction)
+function MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveFunction, func::RiemannianFunction)
     model.objective = func.func
     model.problem = nothing
     model.state = nothing
@@ -351,7 +421,7 @@ end
 # ```
 const DESCENT_STATE_TYPE = "descent_state_type"
 
-function MOI.optimize!(model::Optimizer)
+function MOI.optimize!(model::ManoptOptimizer)
     start = Float64[
         if isnothing(model.variable_primal_start[i])
             error("No starting value specified for `$i`th variable.")
@@ -380,24 +450,48 @@ function MOI.optimize!(model::Optimizer)
     return nothing
 end
 
-struct ArrayShape{N} <: JuMP.AbstractShape
+@doc """
+    ManifoldPointArrayShape{N} <: JuMP.AbstractShape
+
+Represent some generic `AbstractArray` of a certain size representing an point
+on a manifold
+
+# Fields
+
+* `size::NTuple{N,Int}`: The size of the array
+"""
+struct ManifoldPointArrayShape{N} <: JuMP.AbstractShape
     size::NTuple{N,Int}
 end
 
-function JuMP.vectorize(array::Array{T,N}, ::ArrayShape{M}) where {T,N,M}
+"""
+    JuMP.vectorize(p::Array{T,N}, shape::ManifoldPointArrayShape{N}) where {T,N}
+
+Given a point `p` as an ``N``-dimensional array representing a point on a certain
+manifold, reshape it to a vector, which is necessary within [`JuMP`](@extref JuMP :std:doc:`index`).
+For the inverse see [`JuMP.reshape_vector`](@ref JuMP.reshape_vector(::Vector, ::ManifoldPointArrayShape)).
+"""
+function JuMP.vectorize(array::Array{T,N}, ::ManifoldPointArrayShape{M}) where {T,N,M}
     return vec(array)
 end
 
-function JuMP.reshape_vector(vector::Vector, shape::ArrayShape)
+"""
+    JuMP.reshape_vector(vector::Vector, shape::ManifoldPointArrayShape)
+
+Given some vector representation `vector` used within [`JuMP`](@extref JuMP :std:doc:`index`) of a point on a manifold represents points
+by arrays, use the information from the `shape` to reshape it back into such an array.
+For the inverse see [`JuMP.vectorize`](@ref JuMP.vectorize(::Array, ::ManifoldPointArrayShape)).
+"""
+function JuMP.reshape_vector(vector::Vector, shape::ManifoldPointArrayShape)
     return reshape(vector, shape.size)
 end
 
-function JuMP.reshape_set(set::VectorizedManifold, shape::ArrayShape)
+function JuMP.reshape_set(set::ManifoldSet, shape::ManifoldPointArrayShape)
     return set.manifold
 end
 
 function _shape(m::ManifoldsBase.AbstractManifold)
-    return ArrayShape(ManifoldsBase.representation_size(m))
+    return ManifoldPointArrayShape(ManifoldsBase.representation_size(m))
 end
 
 _in(mime::MIME"text/plain") = "in"
@@ -411,26 +505,26 @@ end
     JuMP.build_variable(::Function, func, m::ManifoldsBase.AbstractManifold)
 
 Build a `JuMP.VariablesConstrainedOnCreation` object containing variables
-and the [`Manopt.JuMP_VectorizedManifold`](@ref) in which they should belong as well as the
+and the [`ManifoldSet`](@ref) in which they should belong as well as the
 `shape` that can be used to go from the vectorized MOI representation to the
-shape of the manifold, that is, [`Manopt.JuMP_ArrayShape`](@ref).
+shape of the manifold, that is, [`ManifoldPointArrayShape`](@ref).
 """
 function JuMP.build_variable(::Function, func, m::ManifoldsBase.AbstractManifold)
     shape = _shape(m)
     return JuMP.VariablesConstrainedOnCreation(
-        JuMP.vectorize(func, shape), VectorizedManifold(m), shape
+        JuMP.vectorize(func, shape), ManifoldSet(m), shape
     )
 end
 
 """
-    MOI.get(model::Optimizer, ::MOI.ResultCount)
+    MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)
 
-Return `MOI.OPTIMIZE_NOT_CALLED` if `optimize!` hasn't been called yet and
-`MOI.LOCALLY_SOLVED` otherwise indicating that the solver has solved the
-problem to local optimality the value of `MOI.RawStatusString` for more
+Return [`OPTIMIZE_NOT_CALLED`](@extref `MathOptInterface.OPTIMIZE_NOT_CALLED`) if [`optimize!`](@extref `JuMP.optimize!`) hasn't been called yet and
+[`LOCALLY_SOLVED`](@extref `MathOptInterface.LOCALLY_SOLVED`) otherwise indicating that the solver has solved the
+problem to local optimality the value of [`RawStatusString`](@extref `MathOptInterface.RawStatusString`) for more
 details on why the solver stopped.
 """
-function MOI.get(model::Optimizer, ::MOI.TerminationStatus)
+function MOI.get(model::ManoptOptimizer, ::MOI.TerminationStatus)
     if isnothing(model.state)
         return MOI.OPTIMIZE_NOT_CALLED
     else
@@ -439,12 +533,12 @@ function MOI.get(model::Optimizer, ::MOI.TerminationStatus)
 end
 
 """
-    MOI.get(model::Optimizer, ::MOI.ResultCount)
+    MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)
 
-Return `0` if `optimize!` hasn't been called yet and
+Return `0` if [`optimize!`](@extref `JuMP.optimize!`) hasn't been called yet and
 `1` otherwise indicating that one solution is available.
 """
-function MOI.get(model::Optimizer, ::MOI.ResultCount)
+function MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)
     if isnothing(model.state)
         return 0
     else
@@ -453,13 +547,13 @@ function MOI.get(model::Optimizer, ::MOI.ResultCount)
 end
 
 """
-    MOI.get(model::Optimizer, ::MOI.PrimalStatus)
+    MOI.get(model::ManoptOptimizer, ::MOI.PrimalStatus)
 
-Return `MOI.NO_SOLUTION` if `optimize!` hasn't been called yet and
-`MOI.FEASIBLE_POINT` otherwise indicating that a solution is available
-to query with `MOI.VariablePrimalStart`.
+Return [`MOI.NO_SOLUTION`](@extref JuMP :jl:constant:`MathOptInterface.NO_SOLUTION`) if `optimize!` hasn't been called yet and
+[`MOI.FEASIBLE_POINT`](@extref `MathOptInterface.FEASIBLE_POINT`) if it is otherwise indicating that a solution is available
+to query with [`VariablePrimalStart`](@extref `MathOptInterface.VariablePrimalStart`).
 """
-function MOI.get(model::Optimizer, ::MOI.PrimalStatus)
+function MOI.get(model::ManoptOptimizer, ::MOI.PrimalStatus)
     if isnothing(model.state)
         return MOI.NO_SOLUTION
     else
@@ -468,31 +562,31 @@ function MOI.get(model::Optimizer, ::MOI.PrimalStatus)
 end
 
 """
-    MOI.get(::Optimizer, ::MOI.DualStatus)
+    MOI.get(::ManoptOptimizer, ::MOI.DualStatus)
 
-Returns `MOI.NO_SOLUTION` indicating that there is no dual solution
+Returns [`MOI.NO_SOLUTION`](@extref `MathOptInterface.NO_SOLUTION`) indicating that there is no dual solution
 available.
 """
-MOI.get(::Optimizer, ::MOI.DualStatus) = MOI.NO_SOLUTION
+MOI.get(::ManoptOptimizer, ::MOI.DualStatus) = MOI.NO_SOLUTION
 
 """
-    MOI.get(model::Optimizer, ::MOI.RawStatusString)
+    MOI.get(model::ManoptOptimizer, ::MOI.RawStatusString)
 
-Return a `String` containing `Manopt.get_reason` without the ending newline
+Return a `String` containing [`get_reason`](@ref) without the ending newline
 character.
 """
-function MOI.get(model::Optimizer, ::MOI.RawStatusString)
+function MOI.get(model::ManoptOptimizer, ::MOI.RawStatusString)
     # `strip` removes the `\n` at the end and returns an `AbstractString`
     # Since MOI wants a `String`, pass it through `string`
     return string(strip(get_reason(model.state)))
 end
 
 """
-    MOI.get(model::Optimizer, attr::MOI.ObjectiveValue)
+    MOI.get(model::ManoptOptimizer, attr::MOI.ObjectiveValue)
 
 Return the value of the objective function evaluated at the solution.
 """
-function MOI.get(model::Optimizer, attr::MOI.ObjectiveValue)
+function MOI.get(model::ManoptOptimizer, attr::MOI.ObjectiveValue)
     MOI.check_result_index_bounds(model, attr)
     solution = Manopt.get_solver_return(model.state)
     value = get_cost(model.problem, solution)
@@ -503,11 +597,11 @@ function MOI.get(model::Optimizer, attr::MOI.ObjectiveValue)
 end
 
 """
-    MOI.get(model::Optimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)
+    MOI.get(model::ManoptOptimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)
 
 Return the value of the solution for the variable of index `vi`.
 """
-function MOI.get(model::Optimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)
+function MOI.get(model::ManoptOptimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)
     MOI.check_result_index_bounds(model, attr)
     MOI.throw_if_not_valid(model, vi)
     solution = Manopt.get_solver_return(get_objective(model.problem), model.state)